Avoid @ts-nocheck in package-tests
by using the package.json 'imports' field to map self-imports during
development. We then change the mapped import to a direct import
when running the package tests.

The advantage of avoiding @ts-nocheck is that you can see type errors in
the package test files in real time during development. This is
significant, because the package tests are pretty slow to run.